float32 to float 16

+++ create_logo.py
@@ -28,7 +28,7 @@ def create_pipeline(model_name=model_name_sd): # Sửa tham số mặc định
         print("Using CPU")
         pipeline = StableDiffusionPipeline.from_pretrained(
             model_name,
-            torch_dtype = torch.float32, # Giữ float32 cho CPU
+            torch_dtype = torch.float16, # Giữ float32 cho CPU
             use_safetensors = True
         )
     # Có thể thêm scheduler nếu muốn cải thiện chất lượng/tốc độ (ví dụ DPMSolverMultistepScheduler)
